extract replication method
